[rt.cpan.org #123221] Replace File'Fetch with File::Fetch
In Internals.pm and Internals/Fetch.pm variables in the module File::Fetch are set with the Perl 4 package delimeter File'Fetch. Can you replace File'Fetch with File::Fetch?

I think that the comment in Internals/Fetch.pm only applies to perl 5.0. I've tested CPANPLUS successfully under Perl 5.6.2 as well as current Perl versions with the single quote replaced by a double colon.

I've started to play with cperl, which has removed support for the old delimeter. I've patched CPANPLUS locally, but it would be nice to have this change upstream.

Signed-off-by: Chris 'BinGOs' Williams <chris@bingosnet.co.uk>